+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Compatibility:
 * Implemented `MatchData#dup` (#1792, @XrXr).
 * Implemented a native storage strategy for arrays to allow better C extension compatibility.
 * Implemented `rb_check_symbol_cstr` (#1814).
+* Implemented `rb_hash_start` (#1841, @XrXr).
 
 Performance: